split CloudStorageModel into two files
