+++ client/src/features/project/components/cloudStorage/CloudStorageModal.tsx
@@ -16,43 +16,29 @@
  * limitations under the License.
  */
 
-import { FetchBaseQueryError, skipToken } from "@reduxjs/toolkit/query";
+import { skipToken } from "@reduxjs/toolkit/query";
 import cx from "classnames";
 import { isEqual } from "lodash-es";
 import { useCallback, useEffect, useMemo, useState } from "react";
-import {
-  ArrowCounterclockwise,
-  ArrowRepeat,
-  ChevronLeft,
-  ChevronRight,
-  CloudFill,
-  Database,
-  PencilSquare,
-  PlusLg,
-  XLg,
-} from "react-bootstrap-icons";
-import {
-  Button,
-  Modal,
-  ModalBody,
-  ModalFooter,
-  ModalHeader,
-  UncontrolledTooltip,
-} from "reactstrap";
+import { ArrowCounterclockwise } from "react-bootstrap-icons";
+import { Button, Modal, ModalBody, ModalFooter, ModalHeader } from "reactstrap";
 
-import { SuccessAlert } from "../../../../components/Alert";
-import { Loader } from "../../../../components/Loader";
 import { RtkOrNotebooksError } from "../../../../components/errors/RtkErrorAlert";
+import {
+  CloudStorageGetRead,
+  CloudStoragePatch,
+  PostStoragesV2ApiArg,
+  RCloneConfig,
+  usePatchStoragesV2ByStorageIdMutation,
+  usePostStoragesV2Mutation,
+} from "../../../projectsV2/api/storagesV2.api.ts";
 import {
   findSensitive,
   getCurrentStorageDetails,
   getSchemaProviders,
   hasProviderShortlist,
 } from "../../utils/projectCloudStorage.utils";
-import AddOrEditCloudStorage, {
-  AddOrEditCloudStorageV2,
-  AddStorageBreadcrumbNavbar,
-} from "./AddOrEditCloudStorage";
+import { AddStorageBreadcrumbNavbar } from "./AddOrEditCloudStorage";
 import {
   useAddCloudStorageForProjectMutation,
   useGetCloudStorageSchemaQuery,
@@ -71,299 +57,19 @@ import {
   CloudStorage,
   CloudStorageDetails,
   CloudStorageDetailsOptions,
-  CloudStorageSchema,
   TestCloudStorageConnectionParams,
   UpdateCloudStorageParams,
 } from "./projectCloudStorage.types";
 
 import {
-  CloudStorageGetRead,
-  CloudStoragePatch,
-  PostStoragesV2ApiArg,
-  RCloneConfig,
-  usePatchStoragesV2ByStorageIdMutation,
-  usePostStoragesV2Mutation,
-} from "../../../projectsV2/api/storagesV2.api.ts";
+  AddCloudStorageContinueButton,
+  AddCloudStorageBackButton,
+  AddCloudStorageBodyContent,
+  AddCloudStorageConnectionTestResult,
+  AddCloudStorageHeaderContent,
+} from "./cloudStorageModalComponents.tsx";
+
 import styles from "./CloudStorage.module.scss";
-import { SerializedError } from "@reduxjs/toolkit";
-
-interface AddCloudStorageForwardBackButtonProps {
-  setStateSafe: (newState: Partial<AddCloudStorageState>) => void;
-  state: AddCloudStorageState;
-  validateResult: ReturnType<typeof useTestCloudStorageConnectionMutation>[1];
-}
-
-interface AddCloudStorageBackButtonProps
-  extends AddCloudStorageForwardBackButtonProps {
-  success: boolean;
-  toggle: () => void;
-}
-function AddCloudStorageBackButton({
-  setStateSafe,
-  state,
-  success,
-  toggle,
-  validateResult,
-}: AddCloudStorageBackButtonProps) {
-  if (state.step <= 1 || success)
-    return (
-      <Button
-        className="btn-outline-rk-green"
-        data-cy="cloud-storage-edit-close-button"
-        onClick={() => toggle()}
-      >
-        <XLg className={cx("bi", "me-1")} />
-        {success ? "Close" : "Cancel"}
-      </Button>
-    );
-  return (
-    <Button
-      className="btn-outline-rk-green"
-      data-cy="cloud-storage-edit-back-button"
-      disabled={validateResult.isLoading}
-      onClick={() => {
-        if (!validateResult.isUninitialized) validateResult.reset();
-        setStateSafe({
-          step: state.advancedMode ? 0 : state.step - 1,
-        });
-      }}
-    >
-      <ChevronLeft className={cx("bi", "me-1")} />
-      Back
-    </Button>
-  );
-}
-
-interface AddCloudStorageBodyContentProps
-  extends AddCloudStorageHeaderContentProps {
-  addResultStorageName: string | undefined;
-  redraw: boolean;
-  schema: CloudStorageSchema[] | undefined;
-  schemaError: FetchBaseQueryError | SerializedError | undefined;
-  schemaIsFetching: boolean;
-  setStateSafe: (newState: Partial<AddCloudStorageState>) => void;
-  setStorageDetailsSafe: (
-    newStorageDetails: Partial<CloudStorageDetails>
-  ) => void;
-  state: AddCloudStorageState;
-  storageDetails: CloudStorageDetails;
-  success: boolean;
-}
-function AddCloudStorageBodyContent({
-  addResultStorageName,
-  isV2,
-  redraw,
-  schema,
-  schemaError,
-  schemaIsFetching,
-  setStateSafe,
-  setStorageDetailsSafe,
-  state,
-  storageDetails,
-  storageId,
-  success,
-}: AddCloudStorageBodyContentProps) {
-  if (redraw) return <Loader />;
-  if (success)
-    return (
-      <SuccessAlert dismissible={false} timeout={0}>
-        <p className="p-0">
-          The storage {addResultStorageName} has been successfully{" "}
-          {storageId ? "updated" : "added"}.
-        </p>
-      </SuccessAlert>
-    );
-  if (schemaIsFetching || !schema) return <Loader />;
-  if (schemaError) return <RtkOrNotebooksError error={schemaError} />;
-  if (!isV2) {
-    return (
-      <AddOrEditCloudStorage
-        schema={schema}
-        setState={setStateSafe}
-        setStorage={setStorageDetailsSafe}
-        state={state}
-        storage={storageDetails}
-      />
-    );
-  }
-  return (
-    <>
-      {!storageId && (
-        <p>
-          Add published datasets from data repositories for use in your project.
-          Or, connect to cloud storage to read and write custom data.
-        </p>
-      )}
-      <AddOrEditCloudStorageV2
-        schema={schema}
-        setState={setStateSafe}
-        setStorage={setStorageDetailsSafe}
-        state={state}
-        storage={storageDetails}
-        isV2={isV2}
-      />
-    </>
-  );
-}
-
-interface AddCloudStorageHeaderContentProps {
-  isV2: boolean;
-  storageId: string | null;
-}
-function AddCloudStorageHeaderContent({
-  storageId,
-  isV2,
-}: AddCloudStorageHeaderContentProps) {
-  if (isV2)
-    return (
-      <>
-        <Database className={cx("bi", "me-2")} />{" "}
-        <span className="text-uppercase">
-          {storageId ? "Edit" : "Add"} data source
-        </span>
-      </>
-    );
-  return (
-    <>
-      <CloudFill className={cx("bi", "me-2")} />
-      {storageId ? "Edit" : "Add"} Cloud Storage
-    </>
-  );
-}
-
-interface AddCloudStorageContinueButtonProps
-  extends AddCloudStorageForwardBackButtonProps {
-  addButtonDisableReason: string;
-  addOrEditStorage: () => void;
-  disableAddButton: boolean;
-  disableContinueButton: boolean;
-  isResultLoading: boolean;
-  storageDetails: CloudStorageDetails;
-  storageId: string | null;
-  validateConnection: () => void;
-}
-function AddCloudStorageContinueButton({
-  addButtonDisableReason,
-  addOrEditStorage,
-  disableAddButton,
-  disableContinueButton,
-  isResultLoading,
-  state,
-  setStateSafe,
-  storageDetails,
-  storageId,
-  validateConnection,
-  validateResult,
-}: AddCloudStorageContinueButtonProps) {
-  const addButtonId = "add-cloud-storage-continue";
-  const continueButtonId = "add-cloud-storage-next";
-  if (state.step === 3 && state.completedSteps >= 2) {
-    return (
-      <div id={`${addButtonId}-div`} className="d-inline-block">
-        <Button
-          data-cy="cloud-storage-edit-update-button"
-          id={`${addButtonId}-button`}
-          disabled={disableAddButton}
-          onClick={() => addOrEditStorage()}
-        >
-          {isResultLoading ? (
-            <Loader className="me-1" inline size={16} />
-          ) : storageId ? (
-            <PencilSquare className={cx("bi", "me-1")} />
-          ) : (
-            <PlusLg className={cx("bi", "me-1")} />
-          )}
-          {storageId ? "Update" : "Add"} storage
-        </Button>
-        {disableAddButton && (
-          <UncontrolledTooltip placement="top" target={`${addButtonId}-div`}>
-            {addButtonDisableReason}
-          </UncontrolledTooltip>
-        )}
-      </div>
-    );
-  }
-  if (state.step === 2 && state.completedSteps >= 1) {
-    return (
-      <div id={`${continueButtonId}-div`} className="d-inline-block">
-        <TestConnectionAndContinueButtons
-          actionState={setStateSafe}
-          actionTest={validateConnection}
-          continueId="add-cloud-storage-continue"
-          resetTest={validateResult.reset}
-          step={state.step}
-          testId="test-cloud-storage"
-          testIsFailure={validateResult.isError}
-          testIsOngoing={validateResult.isLoading}
-          testIsSuccess={validateResult.isSuccess}
-        />
-        {disableContinueButton && (
-          <UncontrolledTooltip
-            placement="top"
-            target={`${continueButtonId}-div`}
-          >
-            {!storageDetails.schema
-              ? "Please select a storage type"
-              : "Please select a provider or change storage type"}
-          </UncontrolledTooltip>
-        )}
-      </div>
-    );
-  }
-  return (
-    <div id={`${continueButtonId}-div`} className="d-inline-block">
-      <Button
-        id={`${continueButtonId}-button`}
-        data-cy="cloud-storage-edit-next-button"
-        disabled={disableContinueButton}
-        onClick={() => {
-          setStateSafe({
-            completedSteps:
-              state.step === 0
-                ? CLOUD_STORAGE_TOTAL_STEPS - 1
-                : state.step > state.completedSteps
-                ? state.step
-                : state.completedSteps,
-            step: state.step === 0 ? CLOUD_STORAGE_TOTAL_STEPS : state.step + 1,
-          });
-        }}
-      >
-        Next <ChevronRight className={cx("bi", "me-1")} />
-      </Button>
-      {disableContinueButton && (
-        <UncontrolledTooltip placement="top" target={`${continueButtonId}-div`}>
-          {!storageDetails.schema
-            ? "Please select a storage type"
-            : "Please select a provider or change storage type"}
-        </UncontrolledTooltip>
-      )}
-    </div>
-  );
-}
-
-interface AddCloudStorageConnectionTestResultProps {
-  validateResult: ReturnType<typeof useTestCloudStorageConnectionMutation>[1];
-}
-
-function AddCloudStorageConnectionTestResult({
-  validateResult,
-}: AddCloudStorageConnectionTestResultProps) {
-  if (validateResult.isUninitialized || validateResult.isLoading) return null;
-  if (validateResult.error)
-    return (
-      <div className={cx("w-100", "my-0")}>
-        <RtkOrNotebooksError error={validateResult.error} />
-      </div>
-    );
-  return (
-    <div className={cx("w-100", "my-0")}>
-      {" "}
-      <SuccessAlert timeout={0}>
-        <p className="p-0">The connection to the storage works correctly.</p>
-      </SuccessAlert>
-    </div>
-  );
-}
 
 interface CloudStorageModalProps {
   currentStorage?: CloudStorage | CloudStorageGetRead | null;
@@ -797,117 +503,5 @@ export default function CloudStorageModal({
         )}
       </ModalFooter>
     </Modal>
-  );
-}
-
-interface TestConnectionAndContinueButtonsProps {
-  actionState: (newState: Partial<AddCloudStorageState>) => void;
-  actionTest: () => void;
-  continueId: string;
-  resetTest: () => void;
-  step: number;
-  testId: string;
-  testIsFailure: boolean;
-  testIsOngoing: boolean;
-  testIsSuccess: boolean;
-}
-function TestConnectionAndContinueButtons({
-  actionState,
-  actionTest,
-  continueId,
-  resetTest,
-  step,
-  testId,
-  testIsFailure,
-  testIsOngoing,
-  testIsSuccess,
-}: TestConnectionAndContinueButtonsProps) {
-  const buttonTestId = `${testId}-button`;
-  const divTestId = `${testId}-div`;
-  const testConnectionContent =
-    testIsSuccess || testIsFailure ? (
-      <>
-        Re-test <ArrowRepeat className="bi" />
-      </>
-    ) : testIsOngoing ? (
-      <>
-        Testing connection <Loader inline size={16} />
-      </>
-    ) : (
-      <>
-        Test connection <ChevronRight className={cx("bi", "me-1")} />
-      </>
-    );
-  const testConnectionColorClass = testIsSuccess
-    ? "btn-outline-rk-green"
-    : testIsFailure
-    ? "btn-danger"
-    : "btn-secondary";
-  const testConnectionSection = (
-    <div id={divTestId} className="d-inline-block">
-      <Button
-        color=""
-        id={buttonTestId}
-        data-cy={buttonTestId}
-        className={cx(testConnectionColorClass)}
-        disabled={testIsOngoing}
-        onClick={() => actionTest()}
-      >
-        {testConnectionContent}
-      </Button>
-    </div>
-  );
-
-  const buttonContinueId = `${continueId}-button`;
-  const divContinueId = `${continueId}-div`;
-  const continueContent = testIsSuccess ? (
-    <>
-      Continue <ChevronRight className="bi" />
-    </>
-  ) : testIsFailure ? (
-    <>
-      Skip Test <ChevronRight className="bi" />
-    </>
-  ) : null;
-  const continueColorClass = testIsSuccess
-    ? "btn-secondary"
-    : testIsFailure
-    ? "btn-outline-danger"
-    : "btn-outline-rk-green";
-  const continueSection =
-    !testIsFailure && !testIsSuccess ? null : (
-      <div id={divContinueId} className={cx("d-inline-block", "ms-2")}>
-        <Button
-          color=""
-          id={buttonContinueId}
-          data-cy={buttonContinueId}
-          className={cx(continueColorClass)}
-          disabled={testIsOngoing}
-          onClick={() => {
-            if (testIsFailure || testIsSuccess) {
-              resetTest();
-            }
-            actionState({
-              step: step === 0 ? CLOUD_STORAGE_TOTAL_STEPS : step + 1,
-              completedSteps: step === 0 ? CLOUD_STORAGE_TOTAL_STEPS - 1 : step,
-            });
-          }}
-        >
-          {continueContent}
-        </Button>
-        {testIsFailure && (
-          <UncontrolledTooltip placement="top" target={divContinueId}>
-            The connection is not working as configured. You can make changes
-            and try again, or skip and continue.
-          </UncontrolledTooltip>
-        )}
-      </div>
-    );
-
-  return (
-    <div className="d-inline-block">
-      {testConnectionSection}
-      {continueSection}
-    </div>
   );
 }
